refactor: create 'renterToStringWithEmotion' to centralise emotion CSS logic
