+++ dotcom-rendering/src/web/server/articleToHtml.tsx
@@ -1,8 +1,4 @@
-import createCache from '@emotion/cache';
-import { CacheProvider } from '@emotion/react';
-import createEmotionServer from '@emotion/server/create-instance';
 import { ArticleDesign, ArticlePillar } from '@guardian/libs';
-import { renderToString } from 'react-dom/server';
 import {
 	BUILD_VARIANT,
 	dcrJavascriptBundle,
@@ -26,8 +22,9 @@ import type { TagType } from '../../types/tag';
 import { ArticlePage } from '../components/ArticlePage';
 import { decideFormat } from '../lib/decideFormat';
 import { decideTheme } from '../lib/decideTheme';
+import { renderToStringWithEmotion } from '../lib/emotion';
+import { extractExpeditedIslands } from '../lib/extractIslands';
 import { getHttp3Url } from '../lib/getHttp3Url';
-import { extractExpeditedIslands } from './extractIslands';
 import { pageTemplate } from './pageTemplate';
 import { recipeSchema } from './temporaryRecipeStructuredData';
 
@@ -52,18 +49,9 @@ export const articleToHtml = ({ article }: Props): string => {
 
 	const format: ArticleFormat = decideFormat(article.format);
 
-	const key = 'dcr';
-	const cache = createCache({ key });
-	const emotionServer = createEmotionServer(cache);
-
-	const html = renderToString(
-		<CacheProvider value={cache}>
-			<ArticlePage format={format} CAPIArticle={article} NAV={NAV} />
-		</CacheProvider>,
+	const { html, extractedCss } = renderToStringWithEmotion(
+		<ArticlePage format={format} CAPIArticle={article} NAV={NAV} />,
 	);
-
-	const chunks = emotionServer.extractCriticalToChunks(html);
-	const extractedCss = emotionServer.constructStyleTagsFromChunks(chunks);
 
 	// Expedited islands scripts are added to the document head as 'high priority'
 	const expeditedIslands = extractExpeditedIslands(html);
