+++ dotcom-rendering/src/web/server/blocksToHtml.tsx
@@ -1,9 +1,6 @@
-import createCache from '@emotion/cache';
-import { CacheProvider } from '@emotion/react';
-import createEmotionServer from '@emotion/server/create-instance';
-import { renderToString } from 'react-dom/server';
 import { buildAdTargeting } from '../../lib/ad-targeting';
 import { decideFormat } from '../lib/decideFormat';
+import { renderToStringWithEmotion } from '../lib/emotion';
 import { LiveBlogRenderer } from '../lib/LiveBlogRenderer';
 
 /**
@@ -41,37 +38,28 @@ export const blocksToHtml = ({
 		adUnit,
 	});
 
-	const key = 'dcr';
-	const cache = createCache({ key });
-	const emotionServer = createEmotionServer(cache);
-
-	const html = renderToString(
-		<CacheProvider value={cache}>
-			<LiveBlogRenderer
-				blocks={blocks}
-				format={format}
-				adTargeting={adTargeting}
-				host={host}
-				pageId={pageId}
-				webTitle={webTitle}
-				ajaxUrl={ajaxUrl}
-				isSensitive={isSensitive}
-				isAdFreeUser={isAdFreeUser}
-				switches={switches}
-				isLiveUpdate={true}
-				section={section}
-				// The props below are never used because isLiveUpdate is true but, typescript...
-				shouldHideReaderRevenue={false}
-				tags={[]}
-				isPaidContent={false}
-				contributionsServiceUrl=""
-				keywordIds={keywordIds}
-			/>
-		</CacheProvider>,
+	const { html, extractedCss } = renderToStringWithEmotion(
+		<LiveBlogRenderer
+			blocks={blocks}
+			format={format}
+			adTargeting={adTargeting}
+			host={host}
+			pageId={pageId}
+			webTitle={webTitle}
+			ajaxUrl={ajaxUrl}
+			isSensitive={isSensitive}
+			isAdFreeUser={isAdFreeUser}
+			switches={switches}
+			isLiveUpdate={true}
+			section={section}
+			// The props below are never used because isLiveUpdate is true but, typescript...
+			shouldHideReaderRevenue={false}
+			tags={[]}
+			isPaidContent={false}
+			contributionsServiceUrl=""
+			keywordIds={keywordIds}
+		/>,
 	);
 
-	const chunks = emotionServer.extractCriticalToChunks(html);
-	const extractedCss = emotionServer.constructStyleTagsFromChunks(chunks);
-
-	return extractedCss + html;
+	return `${extractedCss}${html}`;
 };
